Additional/PyInstaller/requirements_CUDA10.2.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
altgraph=0.17=pyhd3eb1b0_0
blas=1.0=mkl
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.6.15=ha878542_0
cairo=1.16.0=h19f5f5c_2
certifi=2022.6.15=py37h89c1867_0
cffi=1.15.0=py37hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=37.0.1=py37h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.1=h6c09931_0
freeglut=3.0.0=hf484d3e_5
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
graphite2=1.3.14=h295c915_1
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py37h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
jasper=2.0.14=hd8c5072_2
jpeg=9e=h7f8727e_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglu=9.0.0=hf484d3e_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lz4-c=1.9.3=h295c915_1
macholib=1.14=pyhd3eb1b0_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
mrcfile=1.3.0=pyh44b312d_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
opencv=3.4.2=py37h6fd60c2_1
openh264=2.1.1=h4ff587b_0
openssl=1.1.1l=h7f98852_0
pcre=8.45=h295c915_0
pillow=9.0.1=py37h22f2fdc_0
pip=21.2.2=py37h06a4308_0
pixman=0.40.0=h7f8727e_1
py-opencv=3.4.2=py37hb342d67_1
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.12.0=py37hdd9d17f_0
pyinstaller=4.8=py37h1ccaba5_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pysocks=1.7.1=py37_1
python=3.7.10=h12debd9_4
python_abi=3.7=2_cp37m
pytorch=1.11.0=py3.7_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
setuptools=61.2.0=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.5=hc218d9a_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.11.0=py37_cu102
torchvision=0.2.1=py37_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=py37h06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
zipp=3.8.0=py37h06a4308_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
